Allow faster polling of templates when the server supports ETags.
When fetching a template send a `If-None-Match` header if we previously
received an ETag, then propogate that out so clients can wait less and
skip extracting image data.

This controlls access to writing the template using a work queue so we
don't do updates in the middle of fetching a template.