Add an overlay when we can't do the minimap.
This started out as trying to do the minimap for arbitrary sites, but
that was going to be a massive amount of work and I'm not actually sure
it's possible. It's still good to abstract out site specific handling,
but the important bit here is the overlay support.

Also turn on the userscript for pxls.space, rplace.tk. and
garlic-bread.reddit.com.