Change ANTONYMPH Fluttershy's ear percing color
We spent quite a while discussing what this pixel even was and analysing
the ANTONYMPH video. In the end we decided that the darker blue wasn't
a good choice as it both didn't match the video, and a lot of people not
using the overlay thought it was vandalism. This patch changes it to a
lighter blue.